Chemkin parser for nasa
